# docs/swagger.yaml
definitions:
  categoryentity.Category:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      updated_at:
        type: string
    type: object
  customerentity.Customer:
    properties:
      address:
        type: string
      customer_id:
        type: integer
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
  entity.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
  /categories:
    post:
      consumes:
      - application/json
      description: Create a new category in the database
      parameters:
      - description: Category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/categoryentity.Category'
      produces:
      - application/json
      responses:
        "201":
          description: Category created
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid category_id format, Unable to parse form data, Unable
            to get Category from form
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application AddCategory error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Add category
      tags:
      - Category
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an Category from the database by ID
      parameters:
      - description: category_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category deleted
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid category_id DeleteCategory
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application DeleteCategory error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Delete category
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: Get Category details by category_id
      parameters:
      - description: category_id
        in: path
        name: category_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get Category
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid category_id GetCategory
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application GetCategory error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Get category
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update a category in the database by ID
      parameters:
      - description: category_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: category updated
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid category_id
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application UpdatCategory error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Update a category
      tags:
      - Category
  /categories/{id}/chain:
    get:
      consumes:
      - application/json
      description: Get category details including all parent categories by category_id
      parameters:
      - description: category_id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get Category chain
          schema:
            $ref: '#/definitions/entity.Response'
        "400":
          description: Invalid category_id GetCategory
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Application GetCategory error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: Get category chain
      tags:
      - Category
...
  /login:
    post:
      consumes:
      - application/json
      description: Log in with user credentials and obtain an access token.
      parameters:
      - description: User credentials for login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/customerentity.Customer'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login
          schema:
            $ref: '#/definitions/entity.Response'
        "422":
          description: Invalid JSON provided
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: User login
      tags:
      - Authorization
  /logout:
    post:
      consumes:
      - application/json
      description: Log out and invalidate the user token.
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful logout
          schema:
            $ref: '#/definitions/entity.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Response'
      summary: User logout
      tags:
      - Authorization
...
swagger: "2.0"